♻️ [Refactoring]: 他ファイルのrate-limit import文を新しい構造に更新
- src/api/client.tsのimport文をrate-limit/に更新
- src/types/errors.tsのimport文をrate-limit/に更新
- src/utils/errors.tsとerrors.test.tsのimport文を新構造に対応
- 新しいフォルダ構造への移行を完了
