+++ src/types/errors.ts
@@ -1,13 +1,10 @@
 // エラー関連の型定義
 
+import type { RateLimitInfo } from '../utils/rate-limit/index.js';
+
 export interface FigmaApiError {
   status: number;
   err: string;
-}
-
-export interface RateLimitInfo {
-  remaining: number;
-  reset: Date;
 }
 
 export interface FigmaError extends Error {
